+++ changelog.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
+⚠️ Breaking changes
+
 - (feat) Recorder and Toolbar can now be extended
+- (feat) Editables group now have its own class
 - (refactor) EditableFactory, Editable and editables
 - (style) Adopt the [Javascript style guide](https://github.com/airbnb/javascript)